+++ Code/Editor/EditorFramework/EditorApp/EditorApp.moc.h
@@ -16,7 +16,7 @@
 #include <Foundation/Types/UniquePtr.h>
 #include <GuiFoundation/ContainerWindow/ContainerWindow.moc.h>
 #include <QApplication>
-#include <ToolsFoundation/Basics/RecentFilesList.h>
+#include <ToolsFoundation/Utilities/RecentFilesList.h>
 #include <ToolsFoundation/Project/ToolsProject.h>
 
 class QMainWindow;
