+++ Code/Tools/Libs/ToolsFoundation/Utilities/RecentFilesList.h
